:recycle: Migrate SpoilerMenu and GateMapDisplay into their own class
